+++ .env.development
@@ -1,0 +1,2 @@
+NODE_ENV=development
+REACT_APP_SERVER_PORT=3456
