+++ src/components/LightControl.scss
@@ -1,0 +1,6 @@
+div.color-picker-container {
+  width: 100%;
+  div.colorPicker {
+    margin: 0;
+  }
+}
